Update DD-PlayerOptions.lua
An unfortunate side effect of removing step statistics in the player options is that if for example switching between single and double a lot, it will remove step statistics when going back to single from double.

Allow it to exist for double despite it not doing anything (for now), but still remove them for other options that don't get changed often (centered notefield/4:3 aspect ratio).