Don't use menu buttons for SongSearch.
The problem with using menu buttons with song search is that if a user has them tied to a keyboard input (like letters/numbers) it will scroll the cursor while typing and lead to unwanted results.

Also added helper text at the bottom so users know to not use their menu buttons and instead use intuitive keyboard/mouse controls.