Change color of column cue if both an arrow and mine are next.
Sometimes when there is a wall of mines during a break before the next note you will see one column blue while the rest are red. Players might think that that arrow is ok to be on when in reality there are still mines before the next arrow.

Let's color these columns a purple color for now to indicate that there are also mines on this column before the note comes.