Fix chart parser for NPS/Measure Counter.
Add a new variable to track NPS separately from how the measure counter works.

If the player has the option to count Jumps/Hands towards the PeakNPS we want to avoid false positives with the measure counter. This ensures that we get both higher density in things with lots of jumps/hands, but also not increase measure counts when there is not stream in places that have jumps/hands.