left/right flip for steps pane display
Having the tabs on the right side of the pane makes having the tabs dynamic be a huge pain. So we'll be flipping the steps info and tabs from left/right to fit better. I will still leave the highscore lists/images as is for now.

Also moved the tech info from density graph into the pane display. It makes more sense to have those things in the same place now.